reorganize mocks and add mocks for software products
